requirements.txt

PyPDF2==3.0.1
faiss-cpu==1.7.4
altair==4
tiktoken>=0.4.0,<0.6.0
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
langchain_openai==0.0.5
python-dotenv==1.0.1
streamlit==1.30.0
bs4==0.0.2
requests

# uncomment to use huggingface llms
# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2